split up project mgmt now two tests, one for project mgmt and one for project rules mgmt
